Update: unit_test template
- add libft
- mod suffix rule
