+++ test/unit_test/template/target.hpp
@@ -6,13 +6,14 @@
 /*   By: dnakano <dnakano@student.42tokyo.jp>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/03/09 10:25:36 by dnakano           #+#    #+#             */
-/*   Updated: 2021/03/09 11:42:24 by dnakano          ###   ########.fr       */
+/*   Updated: 2021/03/16 00:07:30 by dnakano          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #ifndef TARGET_HPP
 #define TARGET_HPP
 
+bool isAlpha(int n);
 bool isOdd(int n);
 bool isEven(int n);
